+++ SMSJournal/users/views.py
@@ -103,68 +103,55 @@ def initialize_journal_prompt(request):
 def initialize_journal(request):
     sub = request.user.subscriber
     if sub.total_entries == 0:
-        print("in if")
         try:
             met = Metrics.objects.get(current=True)
             met.log_journal_entry()
         except:
             pass #never fail user request because of a metrics error
         try:
-            print("in try")
             creds = None
             # The file token.pickle stores the user's access and refresh tokens, and is
             # created automatically when the authorization flow completes for the first
             # time.
-            if os.path.exists(os.path.join(settings.BASE_DIR, str(sub.id) + 'token.pickle')):
-                with open(os.path.join(settings.BASE_DIR, str(sub.id) + 'token.pickle'), 'rb') as token:
-                    creds = pickle.load(token)
+            if settings.DEBUG:
+                if os.path.exists(os.path.join(settings.BASE_DIR, str(sub.id) + 'token.pickle')):
+                    with open(os.path.join(settings.BASE_DIR, str(sub.id) + 'token.pickle'), 'rb') as token:
+                        creds = pickle.load(token)
+            else:
+                pass #EFS call here
             # If there are no (valid) credentials available, let the user log in. ##RIGHT NOW JUST WRITES TO ONE ACCOUNT
-            print("line 119")
-            print(sub.id)
             if not creds or not creds.valid:
                 if creds and creds.expired and creds.refresh_token:
-                    print("creds expired")
                     creds.refresh(Request())
                 else:
-                    print("no creds")
-                    f = open(os.path.join(settings.BASE_DIR, 'credentials.json'), 'w') #THIS IS SO JANKY but we can't keep this as a file in the codebase
-                    f.write(settings.GOOGLE_CREDENTIALS)
-                    f.close()
-                    print("wrote file")
-                    flow = InstalledAppFlow.from_client_secrets_file(
-                        os.path.join(settings.BASE_DIR, 'credentials.json'),
-                        ['https://www.googleapis.com/auth/documents'])
-                    print("made flow")
-                    os.remove(os.path.join(settings.BASE_DIR, 'credentials.json')) #END JANK
-                    print("creds removed")
+                    if settings.DEBUG:
+                        f = open(os.path.join(settings.BASE_DIR, 'credentials.json'), 'w') #THIS IS SO JANKY but we can't keep this as a file in the codebase
+                        f.write(settings.GOOGLE_CREDENTIALS)
+                        f.close()
+                        flow = InstalledAppFlow.from_client_secrets_file(
+                            os.path.join(settings.BASE_DIR, 'credentials.json'),
+                            ['https://www.googleapis.com/auth/documents'])
+                        os.remove(os.path.join(settings.BASE_DIR, 'credentials.json')) #END JANK
+                    else:
+                        pass #EFS call here
                     creds = flow.run_local_server()
-                    print("creds from flow")
                 # Save the credentials for the next run
-                with open(os.path.join(settings.BASE_DIR, str(sub.id) + 'token.pickle'), 'wb') as token:
-                    pickle.dump(creds, token)
-            print("token stuff")
+                if settings.DEBUG:
+                    with open(os.path.join(settings.BASE_DIR, str(sub.id) + 'token.pickle'), 'wb') as token:
+                        pickle.dump(creds, token)
+                else:
+                    pass #EFS call here
             service = build('docs', 'v1', credentials=creds)
-            print("service")
         except:
             return render(request, 'initialize_journal_prompt.html', {"googleError": True})
         names = ["SMSJournal"]
-        print("here")
-        print(names)
         for name in names:
-            print(name)
             try:
-                print("try")
                 journal = sub.journal_set.get(name=process_journal_name(name))
-                print("got journal")
                 write_to_gdoc(journal.google_docs_id, "Welcome to SMSJournal! You can make journal entries to this file by sending messages to the SMSJournal phone number, (970)-507-7992. To send an entry to a different journal, just add a tag like @ideas and we'll find or create the journal \"ideas\" in your Google Drive.", service)
             except:
-                print("except")
                 doc = service.documents().create(body={"title": name}).execute()
-                print("doc made")
-                print(doc["documentId"])
-                print(doc.get("title"))
                 write_to_gdoc(doc["documentId"], "Welcome to SMSJournal! You can make journal entries to this file by sending messages to the SMSJournal phone number, (970)-507-7992. To send an entry to a different journal, just add a tag like @ideas and we'll find or create the journal \"ideas\" in your Google Drive.", service)
-                print("written")
                 journal = Journal(subscriber=sub,
                                   name=process_journal_name(name),
                                   google_docs_id=doc["documentId"])
@@ -204,12 +191,8 @@ def signup_and_charge(request):
 @permission_classes((AllowAny,))
 def stripe_card_change(request):
     try:
-        print("in try")
         sub = User.objects.get(username=request.data["username"]).subscriber
-        print(sub.user.username)
-        print(request.data['stripeToken'])
         sub.change_card(request.data['stripeToken'])
-        print(request.data['stripeToken'])
         return Response({"Result": "Done"})
     except:
         return Response({"Result": "Something went wrong."})
@@ -224,6 +207,3 @@ def unsubscribe(request):
     except:
         pass
     return render(request, 'index.html')
-
-
-
